backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
requests
sqlalchemy
passlib[argon2]
pydantic
python-dotenv 